+++ docs/changelog.rst
@@ -6,6 +6,10 @@ Master
 - TwitchIO
     - Additions
         - Added optional ``started_at`` and ``ended_at`` arguments to :func:`~twitchio.PartialUser.fetch_clips`
+        - Updated docstring regarding new  HypeTrain contribution  method ``OTHER`` for :attr:`~twitchio.HypeTrainContribution.type`
+- ext.eventsub
+    - Additions
+        - Updated docs regarding new HypeTrain contribution method ``other`` for :attr:`~twitchio.ext.eventsub.HypeTrainContributor.type`
 
 2.5.0
 ======